+++ src/index.js
@@ -5,10 +5,15 @@ import {
   createZarrLoader,
   ZarrLoader,
   createOMETiffLoader,
-  OMETiffLoader
+  OMETiffLoader,
+  getChannelStats,
+  OMEZarrReader
 } from './loaders';
+import { DTYPE_VALUES, MAX_SLIDERS_AND_CHANNELS } from './constants';
 
 export {
+  DTYPE_VALUES,
+  MAX_SLIDERS_AND_CHANNELS,
   ScaleBarLayer,
   VivViewerLayer,
   VivViewer,
@@ -16,11 +21,13 @@ export {
   OverviewView,
   PictureInPictureViewer,
   SideBySideView,
+  getChannelStats,
   SideBySideViewer,
   DetailView,
   StaticImageLayer,
   ZarrLoader,
   OMETiffLoader,
   createOMETiffLoader,
-  createZarrLoader
+  createZarrLoader,
+  OMEZarrReader
 };