+++ src/loaders/Pool.js
@@ -9,10 +9,6 @@ const defaultPoolSize =
   typeof navigator !== 'undefined' ? navigator.hardwareConcurrency || 4 : null;
 
 /**
- * @module pool
- */
-
-/**
  * Pool for workers to decode chunks of the images.
  * This is a line-for-line copy of GeoTIFFs old implementation: https://github.com/geotiffjs/geotiff.js/blob/v1.0.0-beta.6/src/pool.js
  */